📦 Moves models to /types
