+++ app/src/lib/stores/socket-store.ts
@@ -1,5 +1,5 @@
 import { writable, type Writable } from 'svelte/store';
-import { type angles } from '$lib/models';
+import { type angles } from '$lib/types/models';
 
 export const servoAnglesOut: Writable<number[]> = writable([
 	0, 45, -90, 0, 45, -90, 0, 45, -90, 0, 45, -90
