+++ app/src/routes/controller/Controls.svelte
@@ -10,7 +10,7 @@
 		type Modes,
 		ModesEnum
 	} from '$lib/stores';
-	import type { vector } from '$lib/models';
+	import type { vector } from '$lib/types/models';
 	import VerticalSlider from '$lib/components/input/vertical-slider.svelte';
 
 	let throttle = new throttler();
